+++ public/term.js
