+++ src/App.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts">
+import Translate from './components/Translate.vue'
+</script>
+
+<template>
+  <Translate />
+</template>
+
+<style scoped>
+</style>
